+++ wp-debugging.php
@@ -3,7 +3,7 @@
  * Plugin Name:       WP Debugging
  * Plugin URI:        https://github.com/afragen/wp-debugging
  * Description:       A support/troubleshooting plugin for WordPress.
- * Version:           1.2.3
+ * Version:           1.2.3.1
  * Author:            Andy Fragen
  * License:           MIT
  * Network:           true
@@ -126,13 +126,22 @@ register_deactivation_hook(
 	}
 );
 
-load_plugin_textdomain( 'wp-debugging' );
+add_action(
+	'init',
+	function() {
+		load_plugin_textdomain( 'wp-debugging' );
+	}
+);
+
 require_once __DIR__ . '/vendor/autoload.php';
 WP_Dependency_Installer::instance()->run( __DIR__ );
 
 add_filter(
-	'wp_dependency_timeout', function( $timeout, $source ) {
+	'wp_dependency_timeout',
+	function( $timeout, $source ) {
 		$timeout = $source !== basename( __DIR__ ) ? $timeout : 14;
 		return $timeout;
-	}, 10, 2
+	},
+	10,
+	2
 );